+++ wasatch/WasatchDevice.py
@@ -18,6 +18,7 @@ from .InterfaceDevice             import InterfaceDevice
 from .BalanceAcquisition          import BalanceAcquisition
 from .SpectrometerState           import SpectrometerState
 from .ControlObject               import ControlObject
+from .AutoRaman                   import AutoRaman
 from .DeviceID                    import DeviceID
 from .Reading                     import Reading
 
@@ -80,6 +81,8 @@ class WasatchDevice(InterfaceDevice):
         self.last_battery_percentage = 0
 
         self.process_f = self._init_process_funcs()
+
+        self.auto_raman = AutoRaman(self)
 
     # ######################################################################## #
     #                                                                          #
@@ -266,32 +269,58 @@ class WasatchDevice(InterfaceDevice):
     # function will "block" while the full 10 measurements are made, and then
     # a single, fully-averaged spectrum will be returned upstream.
     #
-    # @return a Reading object
+    # @return a Reading wrapped in a SpectrometerResponse
     #
     def acquire_spectrum(self):
+        if self.take_one_request and self.take_one_request.auto_raman_request:
+            return self.acquire_spectrum_auto_raman()
+        else:
+            return self.acquire_spectrum_standard()
+
+    def acquire_spectrum_auto_raman(self):
+        """
+        @returns a Reading wrapped in a SpectrometerResponse
+        @todo fold-in a lot of the post-reading sensor measurements 
+              (temperature, interlock etc) provided by acquire_spectrum_standard
+        """
+        log.debug("WasatchDevice.acquire_spectrum_auto_raman: calling AutoRaman.measure")
+        spectrometer_response = self.auto_raman.measure(self.take_one_request.auto_raman_request)
+        reading = spectrometer_response.data
+        log.debug(f"WasatchDevice.acquire_spectrum_auto_raman: received {reading}")
+
+        # return the completed TakeOneRequest and clear our internal handle
+        # @todo self.take_one_request should probably be a list...
+        reading.take_one_request = self.take_one_request
+        self.take_one_request = None
+
+        return spectrometer_response
+
+    def acquire_spectrum_standard(self):
         acquire_response = SpectrometerResponse()
 
         tor = self.take_one_request
-        if tor:
-            log.debug(f"WasatchDevice.acquire_spectrum: attempting to fulfill {tor}")
-        else:
-            log.debug(f"WasatchDevice.acquire_spectrum: no TakeOneRequest in effect")
 
         self.perform_optional_throwaways()
 
         ########################################################################
-        # Batch Collection and Auto-Raman silliness
+        # Batch Collection silliness
         ########################################################################
 
-        # We could move this up into ENLIGHTEN.BatchCollection: have it enable
-        # the laser, wait a bit, and then send the "acquire" command.  But since
-        # WasatchDeviceWrapper.continuous_poll ticks at its own interval, that
-        # would introduce timing interference, and different acquisitions would
-        # realistically end up with different warm-up times for lasers (all "at
-        # least" the configured time, but some longer than others).  Instead,
-        # for now I'm putting this delay here, so it will be exactly the same
-        # (given sleep()'s precision) for each acquisition.  For true precision
-        # this should all go into the firmware anyway.
+        # Note that BatchCollection's "auto-enable laser" is not nearly as
+        # involved as the Auto-Raman performed by acquire_spectrum_auto_raman.
+        # BatchCollection uses TakeOneRequest.enable_laser_before, 
+        # .disable_laser_after, .take_dark, .laser_warmup_ms and .scans_to_average
+        # to take an averaged, dark-corrected Raman measurement WITH FIXED 
+        # ACQUISITION PARAMETERS (integration time, gain). It does not perform 
+        # any optimization of integration time or gain. 
+        #
+        # By implication then, we should simplify this by changing 
+        # BatchCollection to use AutoRamanRequest, adding a checkbox to 
+        # BatchCollection to allow # optimization; if unchecked, min/max integ 
+        # and gain will equal start integ/gain, and no optimization will occur. 
+        # Then we can remove all the auto laser stuff from the following code.
+        #
+        # This has been captured in https://github.com/WasatchPhotonics/ENLIGHTEN/issues/474
 
         auto_enable_laser = tor is not None and tor.enable_laser_before 
         log.debug("acquire_spectrum: auto_enable_laser = %s", auto_enable_laser)